requirements.txt
pandas==2.0.3
numpy==1.24.3
yfinance==0.2.28
newsapi-python==0.2.7
transformers==4.35.2
torch==2.1.0
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
prophet==1.1.5
requests==2.31.0
python-dotenv==1.0.0
jupyter==1.0.0
tqdm==4.66.1
ta==0.10.2